Moved common header file so it gets included in "kernels" so CI runs can see it
